day 12: initial solution
Bit of a messy one, bad repurposing of code.  Part 1 is okay, but for
part 2, here's the rationale: say we have a `waypoint` whose position is
the ship's position and the direction is the `waypoint`'s actual
position.  We just have to `move` the `waypoint`'s direction instead of
the "ship's position".  It's bad from an "abstraction" point of view,
maybe specializing some of the operations per `moveable` would be best,
or just approaching it differently altogether.  Nevertheless, if it
fits, I sits.